Fix failing system tests
