+++ spec/fixtures/files/google_old_places_api_client/autocomplete/edinburgh.json
@@ -1,0 +1,10 @@
+{
+  "predictions": [
+    {
+      "description": "Edinburgh, UK",
+      "place_id": "ChIJdd4hrwug2EcRmSrV3Vo6llI",
+      "types": ["locality", "political"]
+    }
+  ],
+  "status": "OK"
+}
