+++ spec/fixtures/files/google_old_places_api_client/autocomplete/non_existent_city.json
@@ -1,0 +1,4 @@
+{
+  "predictions": [],
+  "status": "OK"
+}
