+++ spec/system/find/results/search_results_subject_and_location_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe 'V2 results - enabled', :js, service: :find do
+RSpec.describe 'Results - enabled', :js, service: :find do
   include FiltersFeatureSpecsHelper
 
   before do
@@ -239,6 +239,23 @@ RSpec.describe 'V2 results - enabled', :js, service: :find do
   end
 
   def when_i_start_typing_an_invalid_location
+    when_i_start_typing_non_existent_city_location
+  end
+
+  def when_i_start_typing_non_existent_city_location
+    stub_request(
+      :get,
+      'https://maps.googleapis.com/maps/api/place/autocomplete/json?components=country:uk&input=NonExistentCity&key=replace_me&language=en&types=geocode'
+    ).with(
+      headers: {
+        'Accept' => '*/*',
+        'Accept-Encoding' => 'gzip;q=1.0,deflate;q=0.6,identity;q=0.3',
+        'Connection' => 'keep-alive',
+        'Keep-Alive' => '30',
+        'User-Agent' => 'Faraday v2.12.2'
+      }
+    ).to_return(status: 200, body: file_fixture('google_old_places_api_client/autocomplete/non_existent_city.json'), headers: { 'Content-Type' => 'application/json' })
+
     fill_in 'City, town or postcode', with: 'NonExistentCity'
   end
